requirements.txt
scrapy==2.11.2
python-dateutil==2.9.0.post0
